rewrite i18n, move files around
