+++ app/components/Utils/Title.tsx
@@ -1,0 +1,3 @@
+export function Title({ children }: { children: React.ReactNode }) {
+  return <div className="mb-5 text-xl font-semibold underline">{children}</div>
+}
