+++ app/i18n/index.ts
@@ -1,39 +1,20 @@
-import { TFunction, createInstance, i18n } from "i18next"
-import resourcesToBackend from "i18next-resources-to-backend"
-import { initReactI18next } from "react-i18next/initReactI18next"
-import { getOptions } from "./settings"
-import { LanguageType } from "@/types/Language"
+import i18next from "./i18n"
+import { headerName } from "./settings"
+import { headers } from "next/headers"
 
-const initI18next = async (lng: LanguageType, ns?: string) => {
-  const i18nInstance = createInstance()
-  await i18nInstance
-    .use(initReactI18next)
-    .use(resourcesToBackend((language: string, namespace: string) => import(`./locales/${language}/${namespace}.json`)))
-    .init(getOptions(lng, ns))
-  return i18nInstance
-}
+export async function useTranslation(ns?: string | string[], options?: Record<string, string>) {
+  const headerList = await headers()
+  const lng = headerList.get(headerName)
 
-export async function useTranslation(
-  lng: LanguageType,
-  ns?: string,
-  options: { keyPrefix?: string } = {},
-): Promise<{ t: TFunction; i18n: i18n }> {
-  const i18nextInstance = await initI18next(lng, ns)
+  if (lng && i18next.resolvedLanguage !== lng) {
+    await i18next.changeLanguage(lng)
+  }
+  if (ns && !i18next.hasLoadedNamespace(ns)) {
+    await i18next.loadNamespaces(ns)
+  }
 
   return {
-    t: i18nextInstance.getFixedT(lng, Array.isArray(ns) ? ns[0] : ns, options.keyPrefix),
-    i18n: i18nextInstance,
+    t: i18next.getFixedT(lng ?? i18next.resolvedLanguage!, Array.isArray(ns) ? ns[0] : ns, options?.keyPrefix),
+    i18n: i18next,
   }
 }
-
-export function getFlagFromLang(country: LanguageType): string {
-  //TODO: improve that
-  const flags = {
-    en: "🇺🇸",
-    fr: "🇫🇷",
-    ko: "🇰🇷",
-  }
-
-  return flags[country]
-}
-
