+++ app/i18n/locales/en/translation.json
@@ -1,58 +1,58 @@
 {
-    "common.en": "English",
-    "common.fr": "French",
-    "common.ko": "Korean",
-    "common.de": "German",
-    "common.yes": "Yes",
-    "common.no": "No",
-    "common.today": "Today",
-    "common.current": "Current",
-    "common.country.ko": "South Korea",
-    "common.country.fr": "France",
-    "common.country.lu": "Luxembourg",
-    "common.city.livange": "Livange",
-    "common.city.contern": "Contern",
-    "common.city.strasbourg": "Strasbourg",
-    "common.city.seoul": "Seoul",
-    "common.city.metz": "Metz",
-    "header.job-title": "Senior Web Developer",
-    "header.download-cv": "Download my CV",
-    "information.title": "Informations",
-    "information.location.key": "Location",
-    "information.experience.key": "Experience",
-    "information.experience.value": "8+ years",
-    "information.remote.key": "Remote work",
-    "information.remote.value": "No/Hybrid",
-    "information.relocation.key": "Relocation",
-    "aboutme.title": "About me",
-    "aboutme.text": "After now many years improving my craft by working on many projects as a consultant or not, I have now enough experience to lead a little team of developers in my current job. I had the chance to work on many different technologies, front or back ones but never stopped getting interest in new emerging technologies. I'm still very specialized in web technologies and even have an official certification from Zend Technologies in PHP. I'm now mostly doing javascript in my work but I'm always open to learn new things.",
-    "education.title": "Education",
-    "education.bac": "Professional Baccalaureate in Electronic Engineering",
-    "education.dut": "BTEC Higher National Diploma, IT",
-    "availability.open": "I'm open for work!",
-    "skills.title": "Skills",
-    "languages.title": "Spoken languages",
-    "experiences.title": "Experiences",
-    "jobtitle.dev": "Web Developer",
-    "jobtitle.dev.junior": "Junior Web Developer",
-    "jobtitle.dev.senior": "Senior Web Developer",
-    "jobtitle.dev.teamlead": "Team Lead Web Developer",
-    "jobtitle.teacher": "Fullstack Web Teacher",
-    "jobdesc.teacher": "Teaching to complete beginner and experienced people CSS, HTML, PHP, Javascript and MySQL",
-    "jobdesc.dev.senior": "Implementation of technical services in legacy applications <br> Migrating a PHP 5 application park in PHP 7 <br> Modernization of a legacy stack <br> Training developers to new tools throughout their implementation <br> Upgrading the development stack with the usage of Docker",
-    "jobdesc.dev.junior": "Development of a payment application using the Symfony 2 framework. <br> Implementation of monitoring of servers and webservices.",
-    "jobdesc.enhans": "Development of multiple web applications and tools oriented mainly around AI tools/chat (GPT-4, Gemini, Claude).",
-    "projects.title": "Personal projects",
-    "project.link": "Link to the project",
-    "project.type.webapp": "Web Application",
-    "project.type.phpframework": "PHP Framework",
-    "project.type.asyncphplib": "Asynchronous PHP Slack client",
-    "project.type.multiwebgame": "Multiplayer Web game",
-    "project.type.restapi": "RESTFUL API",
-    "project.desc.camie": "Complete development of a management application for nurseries",
-    "project.desc.objectivephp": "Development of a PHP micro framework.",
-    "project.desc.slamp": "Development of a slack PHP client using asynchronous features.",
-    "project.desc.jsphotographie": "Personal Photography portfolio written using SvelteKit",
-    "project.desc.neocah": "Adaptation of the card game \"Cards Against Humanity\" as a web game",
-    "project.desc.ffta": "API providing the list of all archery events from the French Archery Federation by scraping the official website"
+  "common.en": "English",
+  "common.fr": "French",
+  "common.ko": "Korean",
+  "common.de": "German",
+  "common.yes": "Yes",
+  "common.no": "No",
+  "common.today": "Today",
+  "common.current": "Current",
+  "common.country.ko": "South Korea",
+  "common.country.fr": "France",
+  "common.country.lu": "Luxembourg",
+  "common.city.livange": "Livange",
+  "common.city.contern": "Contern",
+  "common.city.strasbourg": "Strasbourg",
+  "common.city.seoul": "Seoul",
+  "common.city.metz": "Metz",
+  "header.job-title": "Senior Web Developer",
+  "header.download-cv": "Download my CV",
+  "information.title": "Informations",
+  "information.location.key": "Location",
+  "information.experience.key": "Experience",
+  "information.experience.value": "8+ years",
+  "information.remote.key": "Remote work",
+  "information.remote.value": "No/Hybrid",
+  "information.relocation.key": "Relocation",
+  "aboutme.title": "About me",
+  "aboutme.text": "After now many years improving my craft by working on many projects as a consultant or not, I have now enough experience to lead a little team of developers in my current job. I had the chance to work on many different technologies, front or back ones but never stopped getting interest in new emerging technologies. I'm still very specialized in web technologies and even have an official certification from Zend Technologies in PHP. I'm now mostly doing javascript in my work but I'm always open to learn new things.",
+  "education.title": "Education",
+  "education.bac": "Professional Baccalaureate in Electronic Engineering",
+  "education.dut": "BTEC Higher National Diploma, IT",
+  "availability.open": "I'm open for work!",
+  "skills.title": "Skills",
+  "languages.title": "Spoken languages",
+  "experiences.title": "Experiences",
+  "jobtitle.dev": "Web Developer",
+  "jobtitle.dev.junior": "Junior Web Developer",
+  "jobtitle.dev.senior": "Senior Web Developer",
+  "jobtitle.dev.teamlead": "Team Lead Web Developer",
+  "jobtitle.teacher": "Fullstack Web Teacher",
+  "jobdesc.teacher": "Teaching to complete beginner and experienced people CSS, HTML, PHP, Javascript and MySQL",
+  "jobdesc.dev.senior": "Implementation of technical services in legacy applications <br> Migrating a PHP 5 application park in PHP 7 <br> Modernization of a legacy stack <br> Training developers to new tools throughout their implementation <br> Upgrading the development stack with the usage of Docker",
+  "jobdesc.dev.junior": "Development of a payment application using the Symfony 2 framework. <br> Implementation of monitoring of servers and webservices.",
+  "jobdesc.enhans": "Development of multiple web applications and tools oriented mainly around AI tools/chat (GPT-4, Gemini, Claude).",
+  "projects.title": "Personal projects",
+  "project.link": "Link to the project",
+  "project.type.webapp": "Web Application",
+  "project.type.phpframework": "PHP Framework",
+  "project.type.asyncphplib": "Asynchronous PHP Slack client",
+  "project.type.multiwebgame": "Multiplayer Web game",
+  "project.type.restapi": "RESTFUL API",
+  "project.desc.camie": "Complete development of a management application for nurseries",
+  "project.desc.objectivephp": "Development of a PHP micro framework.",
+  "project.desc.slamp": "Development of a slack PHP client using asynchronous features.",
+  "project.desc.jsphotographie": "Personal Photography portfolio written using SvelteKit",
+  "project.desc.neocah": "Adaptation of the card game \"Cards Against Humanity\" as a web game",
+  "project.desc.ffta": "API providing the list of all archery events from the French Archery Federation by scraping the official website"
 }
