+++ app/i18n/locales/fr/translation.json
@@ -1,58 +1,58 @@
 {
-    "common.en": "Anglais",
-    "common.fr": "Français",
-    "common.ko": "Coréen",
-    "common.de": "Allemand",
-    "common.yes": "Oui",
-    "common.no": "Non",
-    "common.today": "Aujourd'hui",
-    "common.current": "Actuel",
-    "common.country.ko": "Corée du Sud",
-    "common.country.fr": "France",
-    "common.country.lu": "Luxembourg",
-    "common.city.livange": "Livange",
-    "common.city.contern": "Contern",
-    "common.city.strasbourg": "Strasbourg",
-    "common.city.seoul": "Séoul",
-    "common.city.metz": "Metz",
-    "header.job-title": "Développeur Web Senior",
-    "header.download-cv": "Télécharger mon cv",
-    "information.title": "Informations",
-    "information.location.key": "Location",
-    "information.experience.key": "Expérience",
-    "information.experience.value": "8+ ans",
-    "information.remote.key": "Télétravail",
-    "information.remote.value": "Non/Hybride",
-    "information.relocation.key": "Relocation",
-    "aboutme.title": "À propos de moi",
-    "aboutme.text": "Jeune développeur web avec beaucoup d'ambitions, je suis motivé à m'améliorer dans mon domaine. Ayant de l'expérience dans mon travail, je passe régulièrement de mon temps à contribuer à des projets open sources afin d'améliorer mes compétences. Etant passionné par mon travail, je suis de près les technologies émergentes et cherche quotidiennement à améliorer mon expérience et ma connaissance du web.",
-    "education.title": "Éducation",
-    "education.bac": "Baccalauréat STI Génie Électronique",
-    "education.dut": "DUT Informatique",
-    "availability.open": "Je cherche un emploi !",
-    "skills.title": "Compétences",
-    "languages.title": "Langues parlées",
-    "experiences.title": "Expériences",
-    "jobtitle.dev": "Développeur Web",
-    "jobtitle.dev.junior": "Développeur Web Junior",
-    "jobtitle.dev.senior": "Développeur Web Senior",
-    "jobtitle.dev.teamlead": "Team lead développeur Web",
-    "jobtitle.teacher": "Professeur de développement web fullstack",
-    "jobdesc.teacher": "Enseignement à de complet débutants la programmation web avec des outils comme CSS, HTML, PHP, Javascript et MySQL",
-    "jobdesc.dev.senior": "Mise en place de services technique dans une application legacy <br> Migration d'un parc d'application en PHP 5 vers PHP 7 <br> Modernisation d'une stack legacy <br> Entrainement de nouveaux développeurs à l'utilisation de nouveaux outils lors de leurs intégration <br> Mise à jour de la stack de développement avec l'utilisation de Docker",
-    "jobdesc.dev.junior": "Développement d'une application de paiement avec le framework Symfony 2. <br> Mise en place d'un monitoring des serveurs et webservices.",
-    "jobdesc.enhans": "Développement de multiple applications web et outils orienté principalement autours d'outils/chat AI (GPT-4, Gemini, Claude).",
-    "projects.title": "Projets personnels",
-    "project.link": "Lien vers le projet",
-    "project.type.webapp": "Application Web",
-    "project.type.phpframework": "Framework PHP",
-    "project.type.asyncphplib": "Client Slack PHP asynchrone",
-    "project.type.multiwebgame": "Jeu Multijoueur Web",
-    "project.type.restapi": "API RESTFUL",
-    "project.desc.camie": "Développement complet d'une application de gestion pour garderie d'enfants.",
-    "project.desc.objectivephp": "Développement d'un micro framework PHP.",
-    "project.desc.slamp": "Développement d'un client slack PHP utilisant des fonctionnalités asynchrones.",
-    "project.desc.jsphotographie": "Portfolio de photographie personnel utilisant SvelteKit.",
-    "project.desc.neocah": "Adaptation du jeu de carte \"Cards Against Humanity\" en jeu web utilisant firebase.",
-    "project.desc.ffta": "API fournissant la liste des évenements de tir à l'arc à partir de la Fédération Française de tir à l'arc en scrappant le site officiel."
+  "common.en": "Anglais",
+  "common.fr": "Français",
+  "common.ko": "Coréen",
+  "common.de": "Allemand",
+  "common.yes": "Oui",
+  "common.no": "Non",
+  "common.today": "Aujourd'hui",
+  "common.current": "Actuel",
+  "common.country.ko": "Corée du Sud",
+  "common.country.fr": "France",
+  "common.country.lu": "Luxembourg",
+  "common.city.livange": "Livange",
+  "common.city.contern": "Contern",
+  "common.city.strasbourg": "Strasbourg",
+  "common.city.seoul": "Séoul",
+  "common.city.metz": "Metz",
+  "header.job-title": "Développeur Web Senior",
+  "header.download-cv": "Télécharger mon cv",
+  "information.title": "Informations",
+  "information.location.key": "Location",
+  "information.experience.key": "Expérience",
+  "information.experience.value": "8+ ans",
+  "information.remote.key": "Télétravail",
+  "information.remote.value": "Non/Hybride",
+  "information.relocation.key": "Relocation",
+  "aboutme.title": "À propos de moi",
+  "aboutme.text": "Jeune développeur web avec beaucoup d'ambitions, je suis motivé à m'améliorer dans mon domaine. Ayant de l'expérience dans mon travail, je passe régulièrement de mon temps à contribuer à des projets open sources afin d'améliorer mes compétences. Etant passionné par mon travail, je suis de près les technologies émergentes et cherche quotidiennement à améliorer mon expérience et ma connaissance du web.",
+  "education.title": "Éducation",
+  "education.bac": "Baccalauréat STI Génie Électronique",
+  "education.dut": "DUT Informatique",
+  "availability.open": "Je cherche un emploi !",
+  "skills.title": "Compétences",
+  "languages.title": "Langues parlées",
+  "experiences.title": "Expériences",
+  "jobtitle.dev": "Développeur Web",
+  "jobtitle.dev.junior": "Développeur Web Junior",
+  "jobtitle.dev.senior": "Développeur Web Senior",
+  "jobtitle.dev.teamlead": "Team lead développeur Web",
+  "jobtitle.teacher": "Professeur de développement web fullstack",
+  "jobdesc.teacher": "Enseignement à de complet débutants la programmation web avec des outils comme CSS, HTML, PHP, Javascript et MySQL",
+  "jobdesc.dev.senior": "Mise en place de services technique dans une application legacy <br> Migration d'un parc d'application en PHP 5 vers PHP 7 <br> Modernisation d'une stack legacy <br> Entrainement de nouveaux développeurs à l'utilisation de nouveaux outils lors de leurs intégration <br> Mise à jour de la stack de développement avec l'utilisation de Docker",
+  "jobdesc.dev.junior": "Développement d'une application de paiement avec le framework Symfony 2. <br> Mise en place d'un monitoring des serveurs et webservices.",
+  "jobdesc.enhans": "Développement de multiple applications web et outils orienté principalement autours d'outils/chat AI (GPT-4, Gemini, Claude).",
+  "projects.title": "Projets personnels",
+  "project.link": "Lien vers le projet",
+  "project.type.webapp": "Application Web",
+  "project.type.phpframework": "Framework PHP",
+  "project.type.asyncphplib": "Client Slack PHP asynchrone",
+  "project.type.multiwebgame": "Jeu Multijoueur Web",
+  "project.type.restapi": "API RESTFUL",
+  "project.desc.camie": "Développement complet d'une application de gestion pour garderie d'enfants.",
+  "project.desc.objectivephp": "Développement d'un micro framework PHP.",
+  "project.desc.slamp": "Développement d'un client slack PHP utilisant des fonctionnalités asynchrones.",
+  "project.desc.jsphotographie": "Portfolio de photographie personnel utilisant SvelteKit.",
+  "project.desc.neocah": "Adaptation du jeu de carte \"Cards Against Humanity\" en jeu web utilisant firebase.",
+  "project.desc.ffta": "API fournissant la liste des évenements de tir à l'arc à partir de la Fédération Française de tir à l'arc en scrappant le site officiel."
 }
