+++ app/i18n/locales/ko/translation.json
@@ -1,58 +1,58 @@
 {
-    "common.en": "영어",
-    "common.fr": "프랑스어",
-    "common.ko": "한국어",
-    "common.de": "독일어",
-    "common.yes": "네",
-    "common.no": "비선호",
-    "common.today": "오늘",
-    "common.current": "재직중",
-    "common.country.ko": "한국",
-    "common.country.fr": "프랑스",
-    "common.country.lu": "룩셈부르크",
-    "common.city.livange": "Livange",
-    "common.city.contern": "Contern",
-    "common.city.strasbourg": "Strasbourg",
-    "common.city.seoul": "서울",
-    "common.city.metz": "Metz",
-    "header.job-title": "시니어 웹 개발자",
-    "header.download-cv": "이력서 다운로드",
-    "information.title": "정보",
-    "information.location.key": "거주지",
-    "information.experience.key": "경력",
-    "information.experience.value": "8년+",
-    "information.remote.key": "재택근무",
-    "information.remote.value": "비선호",
-    "information.relocation.key": "파견근무",
-    "aboutme.title": "소개",
-    "aboutme.text": "저는 다양한 웹 개발 프로젝트에서 컨설턴트로 참여하거나 독립적으로 작업하면서 풍부한 경험을 쌓아왔습니다. 또한 직전 회사에서 팀 리더로 근무하면서 문제 해결 능력과 소통 능력을 강화하고 충분한 리더십 역량을 갖추게 되었습니다. 경력 기간 동안 프론트엔드와 백엔드에서 다양한 기술을 적용해보았으며, Zend PHP 자격증을 보유하는 등 특히 웹 기술에 탁월한 전문성을 갖추고 있습니다. 현재는 JavaScript를 중점적으로 다루고 있지만, 언제나 새로운 기술에 대한 소식을 확인하고 습득하기 위해 노력하고 있습니다.",
-    "education.title": "학력",
-    "education.bac": "전자 공학 전문 바칼로레아 학사 학위",
-    "education.dut": "BTEC 고등 국가 학위, IT",
-    "availability.open": "구직중",
-    "skills.title": "스킬",
-    "languages.title": "어학 능력",
-    "experiences.title": "경력",
-    "jobtitle.dev": "웹 개발자",
-    "jobtitle.dev.junior": "주니어 웹 개발자",
-    "jobtitle.dev.senior": "시니어 웹 개발자",
-    "jobtitle.dev.teamlead": "웹 개발자 팀장",
-    "jobtitle.teacher": "풀스택 웹 선생님",
-    "jobdesc.teacher": "미래 혹은 현업 개발자에게 CSS, HTML, PHP, Javascript, MySQL 등을 교육",
-    "jobdesc.dev.senior": "레거시 애플리케이션에 기술 서비스 구현 <br> PHP 7에서 PHP 애플리케이션 PARC로 마이그레이션 <br> 레거시 스택의 현대화 <br> 신규 툴 구현을 위한 개발자 교육 <br> Docker를 사용하여 개발 스택 업그레이드 ",
-    "jobdesc.dev.junior": " Symfony 2 프레임워크를 사용한 결제 애플리케이션 개발. <br> 서버 및 웹서비스 모니터링 구현.",
-    "jobdesc.enhans": "AI 도구/채팅을 기반으로 주도하는 여러 웹 애플리케이션과 도구 개발. (GPT-4, Gemini, Claude)",
-    "projects.title": "개인 프로젝트",
-    "project.link": "프로젝트 링크",
-    "project.type.webapp": "웹 앱",
-    "project.type.phpframework": "PHP 프레임워크",
-    "project.type.asyncphplib": "비동기 PHP Slack 클라이언트",
-    "project.type.multiwebgame": "멀티플레이어 웹게임",
-    "project.type.restapi": "RESTFUL API",
-    "project.desc.camie": "어린이집 관리 애플리케이션 개발",
-    "project.desc.objectivephp": "PHP 마이크로 프레임워크 개발",
-    "project.desc.slamp": "비동기 기능을 활용한 PHP Slack 클라이언트 개발",
-    "project.desc.jsphotographie": "SvelteKit를 사용하여 개인 사진 포트폴리오 제작",
-    "project.desc.neocah": "카드 게임, \"Cards Against Humanity\"를 웹 기반 게임으로 적응 개발",
-    "project.desc.ffta": " 프랑스 양궁연맹 공식 홈페이지를 스크랩하여 모든 양궁대회 목록을 제공하는 API 생성"
+  "common.en": "영어",
+  "common.fr": "프랑스어",
+  "common.ko": "한국어",
+  "common.de": "독일어",
+  "common.yes": "네",
+  "common.no": "비선호",
+  "common.today": "오늘",
+  "common.current": "재직중",
+  "common.country.ko": "한국",
+  "common.country.fr": "프랑스",
+  "common.country.lu": "룩셈부르크",
+  "common.city.livange": "Livange",
+  "common.city.contern": "Contern",
+  "common.city.strasbourg": "Strasbourg",
+  "common.city.seoul": "서울",
+  "common.city.metz": "Metz",
+  "header.job-title": "시니어 웹 개발자",
+  "header.download-cv": "이력서 다운로드",
+  "information.title": "정보",
+  "information.location.key": "거주지",
+  "information.experience.key": "경력",
+  "information.experience.value": "8년+",
+  "information.remote.key": "재택근무",
+  "information.remote.value": "비선호",
+  "information.relocation.key": "파견근무",
+  "aboutme.title": "소개",
+  "aboutme.text": "저는 다양한 웹 개발 프로젝트에서 컨설턴트로 참여하거나 독립적으로 작업하면서 풍부한 경험을 쌓아왔습니다. 또한 직전 회사에서 팀 리더로 근무하면서 문제 해결 능력과 소통 능력을 강화하고 충분한 리더십 역량을 갖추게 되었습니다. 경력 기간 동안 프론트엔드와 백엔드에서 다양한 기술을 적용해보았으며, Zend PHP 자격증을 보유하는 등 특히 웹 기술에 탁월한 전문성을 갖추고 있습니다. 현재는 JavaScript를 중점적으로 다루고 있지만, 언제나 새로운 기술에 대한 소식을 확인하고 습득하기 위해 노력하고 있습니다.",
+  "education.title": "학력",
+  "education.bac": "전자 공학 전문 바칼로레아 학사 학위",
+  "education.dut": "BTEC 고등 국가 학위, IT",
+  "availability.open": "구직중",
+  "skills.title": "스킬",
+  "languages.title": "어학 능력",
+  "experiences.title": "경력",
+  "jobtitle.dev": "웹 개발자",
+  "jobtitle.dev.junior": "주니어 웹 개발자",
+  "jobtitle.dev.senior": "시니어 웹 개발자",
+  "jobtitle.dev.teamlead": "웹 개발자 팀장",
+  "jobtitle.teacher": "풀스택 웹 선생님",
+  "jobdesc.teacher": "미래 혹은 현업 개발자에게 CSS, HTML, PHP, Javascript, MySQL 등을 교육",
+  "jobdesc.dev.senior": "레거시 애플리케이션에 기술 서비스 구현 <br> PHP 7에서 PHP 애플리케이션 PARC로 마이그레이션 <br> 레거시 스택의 현대화 <br> 신규 툴 구현을 위한 개발자 교육 <br> Docker를 사용하여 개발 스택 업그레이드 ",
+  "jobdesc.dev.junior": " Symfony 2 프레임워크를 사용한 결제 애플리케이션 개발. <br> 서버 및 웹서비스 모니터링 구현.",
+  "jobdesc.enhans": "AI 도구/채팅을 기반으로 주도하는 여러 웹 애플리케이션과 도구 개발. (GPT-4, Gemini, Claude)",
+  "projects.title": "개인 프로젝트",
+  "project.link": "프로젝트 링크",
+  "project.type.webapp": "웹 앱",
+  "project.type.phpframework": "PHP 프레임워크",
+  "project.type.asyncphplib": "비동기 PHP Slack 클라이언트",
+  "project.type.multiwebgame": "멀티플레이어 웹게임",
+  "project.type.restapi": "RESTFUL API",
+  "project.desc.camie": "어린이집 관리 애플리케이션 개발",
+  "project.desc.objectivephp": "PHP 마이크로 프레임워크 개발",
+  "project.desc.slamp": "비동기 기능을 활용한 PHP Slack 클라이언트 개발",
+  "project.desc.jsphotographie": "SvelteKit를 사용하여 개인 사진 포트폴리오 제작",
+  "project.desc.neocah": "카드 게임, \"Cards Against Humanity\"를 웹 기반 게임으로 적응 개발",
+  "project.desc.ffta": " 프랑스 양궁연맹 공식 홈페이지를 스크랩하여 모든 양궁대회 목록을 제공하는 API 생성"
 }
