--- components/Utils/Title.tsx
@@ -1,5 +1,0 @@
-const Title: React.FC<{ children: React.ReactNode }> = ({ children }) => (
-    <div className="mb-5 text-xl font-semibold underline">{children}</div>
-);
-
-export default Title;
