+++ middleware.ts
@@ -1,45 +1,45 @@
-import { NextRequest, NextResponse } from "next/server";
-import acceptLanguage from "accept-language";
-import { fallbackLng, languages } from "./app/i18n/settings";
+import { NextRequest, NextResponse } from "next/server"
+import acceptLanguage from "accept-language"
+import { fallbackLng, languages, cookieName, headerName } from "@/app/i18n/settings"
 
-acceptLanguage.languages(languages);
+acceptLanguage.languages(languages)
 
 export const config = {
-    // matcher: '/:lng*'
-    matcher: ["/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js).*)"],
-};
-
-const cookieName = "i18next";
+  // Avoid matching for static files, API routes, etc.
+  matcher: ["/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js|site.webmanifest).*)"],
+}
 
 export function middleware(req: NextRequest) {
-    let lng;
-    if (req.cookies.has(cookieName)) {
-        lng = acceptLanguage.get(req.cookies.get(cookieName)?.value);
-    }
-    if (!lng) {
-        lng = acceptLanguage.get(req.headers.get("Accept-Language"));
-    }
-    if (!lng) {
-        lng = fallbackLng;
-    }
+  // Ignore paths with "icon" or "chrome"
+  if (req.nextUrl.pathname.indexOf("icon") > -1 || req.nextUrl.pathname.indexOf("chrome") > -1)
+    return NextResponse.next()
 
-    // Redirect if lng in path is not supported
-    if (
-        !languages.some(loc => req.nextUrl.pathname.startsWith(`/${loc}`)) &&
-        !req.nextUrl.pathname.startsWith("/_next")
-    ) {
-        return NextResponse.redirect(new URL(`/${lng}${req.nextUrl.pathname}`, req.url));
-    }
+  let lng
+  // Try to get language from cookie
+  if (req.cookies.has(cookieName)) lng = acceptLanguage.get(req.cookies.get(cookieName)?.value)
+  // If no cookie, check the Accept-Language header
+  if (!lng) lng = acceptLanguage.get(req.headers.get("Accept-Language"))
+  // Default to fallback language if still undefined
+  if (!lng) lng = fallbackLng
 
-    if (req.headers.has("referer")) {
-        const refererUrl = new URL(req.headers.get("referer") || "");
-        const lngInReferer = languages.find((l: string) => refererUrl.pathname.startsWith(`/${l}`));
-        const response = NextResponse.next();
-        if (lngInReferer) {
-            response.cookies.set(cookieName, lngInReferer);
-        }
-        return response;
-    }
+  // Check if the language is already in the path
+  const lngInPath = languages.find((loc) => req.nextUrl.pathname.startsWith(`/${loc}`))
+  const headers = new Headers(req.headers)
+  headers.set(headerName, lngInPath || lng)
 
-    return NextResponse.next();
+  // If the language is not in the path, redirect to include it
+  if (!lngInPath && !req.nextUrl.pathname.startsWith("/_next")) {
+    return NextResponse.redirect(new URL(`/${lng}${req.nextUrl.pathname}${req.nextUrl.search}`, req.url))
+  }
+
+  // If a referer exists, try to detect the language from there and set the cookie accordingly
+  if (req.headers.has("referer")) {
+    const refererUrl = new URL(req.headers.get("referer") || "", req.url)
+    const lngInReferer = languages.find((l) => refererUrl.pathname.startsWith(`/${l}`))
+    const response = NextResponse.next({ headers })
+    if (lngInReferer) response.cookies.set(cookieName, lngInReferer)
+    return response
+  }
+
+  return NextResponse.next({ headers })
 }
