+++ package.json
@@ -12,20 +12,17 @@
     "@vercel/analytics": "^1.5.0",
     "@vercel/speed-insights": "^1.2.0",
     "accept-language": "^3.0.20",
-    "autoprefixer": "10.4.21",
-    "client-only": "^0.0.1",
     "clsx": "^2.1.1",
     "i18next": "^25.3.2",
     "i18next-browser-languagedetector": "^8.2.0",
     "i18next-resources-to-backend": "^1.2.1",
     "js-cookie": "^3.0.5",
+    "lucide-react": "^0.525.0",
     "next": "^15.4.1",
     "next-themes": "^0.4.6",
-    "postcss": "8.5.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-i18next": "^15.6.0",
-    "react-icons": "^5.5.0",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
@@ -39,6 +36,7 @@
     "eslint": "^9.31.0",
     "eslint-config-next": "15.4.1",
     "eslint-config-prettier": "^10.1.5",
+    "postcss": "8.5.6",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.6.14",
     "tailwindcss": "^4.1.11",
